# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Short Link API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/openapis/": {
            "get": {
                "description": "短链接跳转",
                "tags": [
                    "短链接跳转"
                ],
                "summary": "短链接跳转",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接",
                        "name": "shortlink",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/openapis/hello": {
            "get": {
                "description": "健康检查",
                "tags": [
                    "短链接健康检查"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/openapis/short": {
            "post": {
                "description": "获取锻链接",
                "tags": [
                    "获取锻链接"
                ],
                "summary": "获取锻链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "原始的长链接",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 7
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "successed"
                }
            }
        }
    }
}